pyproject.toml
[project]
name = "sibyl"
version = "0.1.0"
description = "AutoMod and Toxicity Profiles using ML"
readme = "README.md"
license = "MPL-2.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "Framework :: Django",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)"
]
requires-python = ">=3.13"
dependencies = [
    "dj-database-url",
    "django",
    "djangorestframework",
    "drf-spectacular",
    "loguru",
    "psycopg[binary]",
    "sentry-sdk[django]",
    "uvicorn",
    "whitenoise",
]

[project.urls]
homepage = "https://hakase.dragonejt.dev"
source = "https://github.com/dragonejt/hakase.git"
issues = "https://github.com/dragonejt/hakase/issues"
documentation = "https://github.com/dragonejt/hakase/wiki"

[dependency-groups]
dev = [
    "coverage",
    "ruff",
]

[tool.ruff]
line-length = 100
target-version = "py313"
extend-exclude = [
    "**/migrations/*",
    "manage.py"
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "PL",  # Pylint
    "I",   # isort
    "DJ",  # Django
    "C90", # mccabe
]